Split into builder and transition
