+++ dpc/src/record/encrypted.rs
@@ -91,11 +91,11 @@ impl<C: Parameters> EncryptedRecord<C> {
             "The DPC assumes that the record is less than 65535 bytes."
         );
 
-        // Encrypt the record plaintext
-        let record_public_key = record.owner().to_encryption_key();
-        let encryption_randomness = C::account_encryption_scheme().generate_randomness(record_public_key, rng)?;
+        // Encrypt the record plaintext.
+        let encryption_key = record.owner().to_encryption_key();
+        let encryption_randomness = C::account_encryption_scheme().generate_randomness(encryption_key, rng)?;
         let encrypted_record =
-            C::account_encryption_scheme().encrypt(record_public_key, &encryption_randomness, &bytes)?;
+            C::account_encryption_scheme().encrypt(encryption_key, &encryption_randomness, &bytes)?;
         let encrypted_record = Self::new(encrypted_record);
 
         Ok((encrypted_record, encryption_randomness))
